+++ src/App.tsx
@@ -24,7 +24,7 @@ import '@/App.css';
 
 import { ToastProvider } from '@/ui/components/common/ToastProvider';
 import SseListener from '@/ui/components/common/SseListener';
-import FeedSearch from './components/features/feed/search/FeedSearch';
+import FeedAndMemberSearch from './components/features/feed/search/FeedAndMemberSearch';
 
 function App() {
     return (
@@ -62,7 +62,7 @@ function App() {
                     path="/search"
                     element={
                         <ProtectedRoute>
-                            <FeedSearch />
+                            <FeedAndMemberSearch />
                         </ProtectedRoute>
                     }
                 />
@@ -84,6 +84,14 @@ function App() {
                 />
                 <Route
                     path="/profile"
+                    element={
+                        <ProtectedRoute>
+                            <Profile />
+                        </ProtectedRoute>
+                    }
+                />
+                <Route
+                    path="/profile/:accountId"
                     element={
                         <ProtectedRoute>
                             <Profile />
